+++ test/test_widgets/lib/dynamic_flow_adapter/dynamic_flow_adapter.dart
@@ -8,7 +8,7 @@ import '../test_base.dart';
 bool toDoListEffect(Action action, Context<ToDoList> ctx) {
   if (action.type == ToDoListAction.onAdd) {
     print('adapter onAdd');
-    ctx.dispatch(Action(ToDoListAction.add, payload: ToDo.mock()));
+    ctx.dispatch(Action(ToDoListAction.add, payload: Todo.mock()));
     return true;
   }
 
@@ -17,13 +17,13 @@ bool toDoListEffect(Action action, Context<ToDoList> ctx) {
 
 ToDoList toDoListReducer(ToDoList state, Action action) {
   print('onReduce:${action.type}');
-  if (!(action.payload is ToDo)) return state;
+  if (!(action.payload is Todo)) return state;
 
   if (action.type == ToDoListAction.add) {
     return state.clone()..list.add(action.payload);
   } else if (action.type == ToDoListAction.remove) {
     return state.clone()
-      ..list.removeWhere((ToDo toDo) => toDo.id == action.payload.id);
+      ..list.removeWhere((Todo toDo) => toDo.id == action.payload.id);
   } else {
     return state.clone();
   }
@@ -33,12 +33,12 @@ final testAdapter = TestDynamicFlowAdapter<ToDoList>(
     pool: <String, ToDoComponent>{'toDo': ToDoComponent()},
     connector: Connector<ToDoList, List<ItemBean>>(
         get: (ToDoList toDoList) => toDoList.list
-            .map<ItemBean>((ToDo toDo) => ItemBean('toDo', toDo))
+            .map<ItemBean>((Todo toDo) => ItemBean('toDo', toDo))
             .toList(),
         set: (ToDoList toDoList, List<ItemBean> beans) {
           toDoList.list.clear();
           toDoList.list
-              .addAll(beans.map<ToDo>((ItemBean bean) => bean.data).toList());
+              .addAll(beans.map<Todo>((ItemBean bean) => bean.data).toList());
         }),
     reducer: toDoListReducer,
     effect: toDoListEffect);