+++ CHANGELOG.md
@@ -1,3 +1,13 @@
-## [0.0.1] - TODO: Add release date.
+## [0.1.1]
 
-* TODO: Describe initial release.
+-   fix hot-reload bug
+-   add excluedSelf in broadcast
+-   rename sample to example
+-   rename docs to doc
+
+## [0.1.2]
+
+-   add stfState field in Context #58
+-   add batchedNotify feature in page-store
+-   add some docs
+-   fix some misspellings